hjaaj. Fixed error in TR1FCK for MC-srDFT properties.
All ground state MC-srDFT optimizations were not affected by this error,
excited state MC-srDFT could be, and srdft-molhes (under development)
certainly was affected, it was because of obviously wrong molhes
elements that the error was discovered.

The fix:

-         CALL DZERO(FXCAO,NDMAT*N2BASX)
+         IF (NDTMAT .GT. 0)
+     &   CALL DZERO(FTVAO,NDTMAT*N2BASX)
+         IF (NDXMAT .GT. 0)
+     &   CALL DZERO(FXCAO,NDXMAT*N2BASX)

The "-" line would not zero all of FTVAO if both NDTMAT and NDXMAT > 0.

And a little polishing elsewhere.